file_operation.py: clean up code in makemain()
The main.py will be now generated in a more "cleaner" manner. Besides, we don't need to write the date of creation to the main.py, as that can be -- if that's what user wants to do -- denoted in the form of a dunder meta variable in the main.py, like '__created__ = "January 1 2022"', for example.

Previously, the author info, like author's name, was automatically written to the main.py. It was done by grabbing user's git config and then what ever user had specified in the name -field, that was used. However, this method seems a little unstable and useless, as user will edit the main.py anyway and might just denote the author info there by hand.